@@ -15,33 +15,77 @@
 //  limitations under the License.
 
 import React from "react";
-import { WithContext as ReactTags } from 'react-tag-input';
+import CreateReactClass from 'create-react-class';
+import PropTypes from 'prop-types';
+import Select from 'react-select';
 import { getUrlPrefix } from "jsx/mixin";
 import $ from "jquery";
+import DatePicker from "react-datepicker";
+import Moment from 'moment';
+import moment from "moment";
 
+Moment.locale(navigator.language);
 
-const repairForm = React.createClass({
+const repairForm = CreateReactClass({
 
   propTypes: {
-    addRepairSubject: React.PropTypes.object.isRequired,
-    addRepairResult: React.PropTypes.object.isRequired,
-    clusterNames: React.PropTypes.object.isRequired
+    addRepairSubject: PropTypes.object.isRequired,
+    addRepairResult: PropTypes.object.isRequired,
+    clusterNames: PropTypes.object.isRequired,
+    currentCluster: PropTypes.string.isRequired,
+    formType: PropTypes.string.isRequired,
   },
 
   getInitialState: function() {
     const URL_PREFIX = getUrlPrefix(window.top.location.pathname);
-    
+
     return {
-      addRepairResultMsg: null, clusterNames: [], submitEnabled: false,
-      clusterName: this.props.currentCluster!="all"?this.props.currentCluster:this.props.clusterNames[0], keyspace: "", tables: "", owner: null, segments: null,
-      parallelism: null, intensity: null, cause: null, incrementalRepair: null, formCollapsed: true, nodes: "", datacenters: "", blacklistedTables: "",
-      nodeList: [], datacenterList: [], clusterStatus: {}, urlPrefix: URL_PREFIX, nodeSuggestions: [], datacenterSuggestions: [], tableSuggestions: [], 
-      clusterTables: {}, blacklistSuggestions: [], tableList: [], blacklistList: [], keyspaceList: [], keyspaceSuggestions: [],
-      blacklistReadOnly: false, tablelistReadOnly: false, advancedFormCollapsed: true, repairThreadCount: 1
+      formType: this.props.formType,
+      addRepairResultMsg: null,
+      clusterNames: [],
+      submitEnabled: false,
+      clusterName: this.props.currentCluster === "all" ? this.props.clusterNames[0] : this.props.currentCluster,
+      keyspace: "",
+      tables: "",
+      blacklistedTables: "",
+      nodes: "",
+      datacenters: "",
+      owner: "",
+      segments: "",
+      parallelism: null,
+      intensity: undefined,
+      cause: "",
+      startTime: this.props.formType === "schedule" ? moment().toDate() : null,
+      intervalDays: "",
+      incrementalRepair: "false",
+      repairThreadCount: 1,
+      formCollapsed: true,
+      advancedFormCollapsed: true,
+      clusterStatus: {},
+      clusterTables: {},
+      urlPrefix: URL_PREFIX,
+      keyspaceOptions: [],
+      tableOptions: [],
+      nodeOptions: [],
+      datacenterOptions: [],
+      keyspaceList: [],
+      keyspaceSelectValues: [],
+      tablesList: [],
+      tablesSelectValues: [],
+      tablesSelectDisabled: false,
+      blacklistedTablesList: [],
+      blacklistedTablesSelectValues: [],
+      blacklistedTablesSelectDisabled: false,
+      nodesList: [],
+      nodesSelectValues: [],
+      nodesSelectDisabled: false,
+      datacentersList: [],
+      datacentersSelectValues: [],
+      datacentersSelectDisabled: false,
     };
   },
 
-  componentWillMount: function() {
+  UNSAFE_componentWillMount: function() {
     this._repairResultSubscription = this.props.addRepairResult.subscribeOnNext(obs =>
       obs.subscribe(
         r => this.setState({addRepairResultMsg: null}),
@@ -53,8 +97,10 @@ const repairForm = React.createClass({
       obs.subscribeOnNext(names => {
         let previousNames = this.state.clusterNames;
         this.setState({clusterNames: names});
-        if(names.length == 1) this.setState({clusterName: names[0]});
-        if(previousNames.length == 0) {
+        if (names.length) {
+          this.setState({clusterName: names[0]});
+        }
+        if (!previousNames.length) {
           this._getClusterStatus();
         }
       })
@@ -66,93 +112,125 @@ const repairForm = React.createClass({
     this._clusterNamesSubscription.dispose();
   },
 
-
   _getClusterStatus: function() {
     let clusterName = this.state.clusterName;
 
-    if (clusterName !== null) {
+    if (clusterName) {
       $.ajax({
-          url: this.state.urlPrefix + '/cluster/' + encodeURIComponent(clusterName),
-          method: 'GET',
-          component: this,
-          complete: function(data) {
-            this.component.setState({clusterStatus: $.parseJSON(data.responseText)});
-            this.component._getSuggestions();
-          }
+        url: this.state.urlPrefix + '/cluster/' + encodeURIComponent(clusterName),
+        method: 'GET',
+        component: this,
+        complete: function(data) {
+          this.component.setState({clusterStatus: $.parseJSON(data.responseText)});
+          this.component._getDatacenterOptions();
+          this.component._getNodeOptions();
+        }
       });
-        $.ajax({
+      $.ajax({
         url: this.state.urlPrefix + '/cluster/' + encodeURIComponent(clusterName) + '/tables',
         method: 'GET',
         component: this,
         complete: function(data) {
           this.component.setState({clusterTables: $.parseJSON(data.responseText)});
-          this.component._getKeyspaceSuggestions();
+          this.component._getKeyspaceOptions();
         }
       });
     }
   },
 
-  _getSuggestions: function() {
-    var nodes = this.state.clusterStatus.nodes_status.endpointStates[0].endpointNames;
-    nodes.sort();
-    this.state.nodeSuggestions = nodes;
-    
-    var datacenters = Object.keys(this.state.clusterStatus.nodes_status.endpointStates[0].endpoints);
-    datacenters.sort;
-    this.setState({datacenterSuggestions: datacenters});
+  _getDatacenterOptions: function() {
+    this.setState({
+      datacenterOptions: Object.keys(this.state.clusterStatus.nodes_status.endpointStates[0].endpoints).sort().map(
+        obj => { return {value: obj, label: obj}; }
+      )
+    });
   },
 
-  _getKeyspaceSuggestions: function() {
-    this.setState({keyspaceSuggestions: Object.keys(this.state.clusterTables)});
+  _getNodeOptions: function() {
+    this.setState({
+      nodeOptions: this.state.clusterStatus.nodes_status.endpointStates[0].endpointNames.sort().map(
+        obj => { return {value: obj, label: obj}; }
+      )
+    });
   },
 
-  _getTableSuggestions: function(ks) {
-    this.setState({tableSuggestions: this.state.clusterTables[ks]});
+  _getKeyspaceOptions: function() {
+    this.setState({keyspaceOptions: Object.keys(this.state.clusterTables).map(
+      obj => { return {value: obj, label: obj}; }
+    )});
+  },
+
+  _getTableOptions: function(ks) {
+    let tableOptionsRef = [];
+    if (ks) {
+      tableOptionsRef = this.state.clusterTables[ks].map(
+        obj => { return {value: obj, label: obj}; }
+      );
+    }
+
+    this.setState({tableOptions: tableOptionsRef});
   },
 
   _onAdd: function(e) {
-    const repair = {
-      clusterName: this.state.clusterName, keyspace: this.state.keyspace,
+    let repair = {
+      clusterName: this.state.clusterName,
+      keyspace: this.state.keyspace,
       owner: this.state.owner
     };
-    if(this.state.tables) repair.tables = this.state.tables;
-    if(this.state.segments) repair.segmentCount = this.state.segments;
-    if(this.state.parallelism) repair.repairParallelism = this.state.parallelism;
-    if(this.state.intensity) repair.intensity = this.state.intensity;
-    if(this.state.cause) repair.cause = this.state.cause;
-    if(this.state.incrementalRepair) repair.incrementalRepair = this.state.incrementalRepair;
-    if(this.state.nodes) repair.nodes = this.state.nodes;
-    if(this.state.datacenters) repair.datacenters = this.state.datacenters;
-    if(this.state.blacklistedTables) repair.blacklistedTables = this.state.blacklistedTables;
-    if(this.state.repairThreadCount && this.state.repairThreadCount > 0) repair.repairThreadCount = this.state.repairThreadCount;
 
-    // Force incremental repair to FALSE if empty
-    if(!this.state.incrementalRepair) repair.incrementalRepair = "false";
+    if (this.state.formType === "schedule") {
+      repair["scheduleTriggerTime"] = moment(this.state.startTime).utc().format("YYYY-MM-DDTHH:mm");
+      repair["scheduleDaysBetween"] = this.state.intervalDays;
+    }
 
-    this.props.addRepairSubject.onNext(repair);
+    if (this.state.tables) repair.tables = this.state.tables;
+    if (this.state.segments) repair.segmentCount = this.state.segments;
+    if (this.state.parallelism) repair.repairParallelism = this.state.parallelism;
+    if (this.state.intensity) repair.intensity = this.state.intensity;
+    if (this.state.cause) repair.cause = this.state.cause;
+    if (this.state.incrementalRepair) {
+      repair.incrementalRepair = this.state.incrementalRepair;
+    }
+    else {
+      repair.incrementalRepair = "false";
+    }
+    if (this.state.nodes) repair.nodes = this.state.nodes;
+    if (this.state.datacenters) repair.datacenters = this.state.datacenters;
+    if (this.state.blacklistedTables) repair.blacklistedTables = this.state.blacklistedTables;
+    if (this.state.repairThreadCount && this.state.repairThreadCount > 0) repair.repairThreadCount = this.state.repairThreadCount;
+
+    this.props.addRepairSubject.onNext({
+      type: this.state.formType,
+      params: repair,
+    });
   },
 
   _handleChange: function(e) {
     var v = e.target.value;
-    var n = e.target.id.substring(3); // strip in_ prefix
+    var n = e.target.id.split("_")[1];
 
     // update state
     const state = this.state;
     state[n] = v;
     this.replaceState(state);
-    
-    if (n == 'clusterName') {
-      this._getClusterStatus();
-    }
-    
+
     // validate
     this._checkValidity();
   },
 
   _checkValidity: function() {
-    const valid = this.state.keyspaceList.length > 0 && this.state.clusterName && this.state.owner
-                                      && ((this.state.datacenterList.length>0 && this.state.nodeList.length==0)
-                                      || (this.state.datacenterList.length==0 && this.state.nodeList.length > 0) || (this.state.datacenterList.length==0  && this.state.nodeList==0) );
+    const valid = this.state.keyspaceList.length
+      && this.state.clusterName
+      && this.state.owner
+      && (
+        (this.state.formType === "schedule" && this.state.startTime && this.state.intervalDays)
+        || (this.state.formType === "repair" && !this.state.startTime && !this.state.intervalDays)
+      )
+      && (
+        (this.state.datacentersList.length && !this.state.nodesList.length)
+          || (!this.state.datacentersList.length && this.state.nodesList.length)
+          || (!this.state.datacentersList.length && !this.state.nodesList.length)
+      );
     this.setState({submitEnabled: valid});
   },
 
@@ -174,167 +252,110 @@ const repairForm = React.createClass({
     }
   },
 
-  _handleAddition(node) {
-    if (this.state.datacenterList.length == 0 && node.length > 1) {
-      let nodes = this.state.nodeList;
-      if ($.inArray(node, this.state.nodes.split(','))==-1) {
-        nodes.push({
-            id: nodes.length + 1,
-            text: node
-        });
-        this.setState({nodeList: nodes, nodes: nodes.map(node => node.text).join(',')});
-        this._checkValidity();
-      }
+  _handleSelectOnChange: function(valueContext, actionContext) {
+    const stateName = actionContext.name.split("_")[1];
+    let stateValue = {};
+
+    stateValue[stateName] = valueContext.value;
+    this.setState(stateValue);
+
+    if (stateName === "clusterName") {
+      this._getClusterStatus();
     }
-  },
-  
-  _handleDelete(i) {
-        let nodes = this.state.nodeList;
-        nodes.splice(i, 1);
-        this.setState({nodeList: nodes, nodes: nodes.map(node => node.text).join(',')});
-        this._checkValidity();
+
+    this._checkValidity();
   },
 
-  _handleDcAddition(dc) {
-    if (this.state.nodeList.length == 0 && dc.length > 1) {
-      let datacenters = this.state.datacenterList;
-      if ($.inArray(dc, this.state.datacenters.split(','))==-1) {
-        datacenters.push({
-            id: datacenters.length + 1,
-            text: dc
-        });
-        this.setState({datacenterList: datacenters, datacenters: datacenters.map(dc => dc.text).join(',')});
-        this._checkValidity();
-      }
+  _getOppositeSelect: function(selectName) {
+    const selectNameMap = {
+      "tables": "blacklistedTables",
+      "blacklistedTables": "tables",
+      "nodes": "datacenters",
+      "datacenters": "nodes",
+    };
+    return selectNameMap[selectName];
+  },
+
+  _handleLinkedSelectOnChange: function(valueContext, actionContext) {
+    const nameRef = actionContext.name.split("_")[1];
+    const oppositeNameRef = this._getOppositeSelect(nameRef);
+    const oppositeNameListRef = `${oppositeNameRef}List`;
+
+    if (this.state[oppositeNameListRef].length) {
+      return;
     }
+
+    const nameListRef = `${nameRef}List`;
+    const nameSelectValuesRef = `${nameRef}SelectValues`;
+    const oppositeNameDisabledRef = `${oppositeNameRef}SelectDisabled`;
+
+    let newList = this.state[nameListRef];
+    let newSelectValues = [];
+    let newRef = "";
+
+    newList.length = 0;
+
+    if (valueContext) {
+      newList = valueContext.map(
+        obj => { return {id: this._create_UUID(), text: obj.value}; }
+      );
+      newRef = valueContext.map(
+        obj => { return obj.value; }
+      ).join(",");
+      newSelectValues = valueContext.map(
+        obj => { return {label: obj.value, value: obj.value}; }
+      );
+    }
+
+    let newState = {};
+    newState[nameRef] = newRef;
+    newState[nameListRef] = newList;
+    newState[nameSelectValuesRef] = newSelectValues;
+    newState[oppositeNameDisabledRef] = newList.length > 0;
+
+    this.setState(newState);
+    this._checkValidity();
   },
-  
-  _handleDcDelete(i) {
-        let datacenters = this.state.datacenterList;
-        datacenters.splice(i, 1);
-        this.setState({datacenterList: datacenters, datacenters: datacenters.map(dc => dc.text).join(',')});
-        this._checkValidity();
-  },
 
-  _handleNodeFilterSuggestions(textInputValue, possibleSuggestionsArray) {
-    var lowerCaseQuery = textInputValue.toLowerCase();
-    let nodes = this.state.nodes;
- 
-    return possibleSuggestionsArray.filter(function(suggestion)  {
-        return suggestion.toLowerCase().includes(lowerCaseQuery) && $.inArray(suggestion, nodes.split(','))==-1;
-    })
-  },
+  // Keyspace list functions
+  _handleKeyspaceSelectOnChange: function(valueContext, actionContext) {
+    let keyspaceListRef = this.state.keyspaceList;
+    let keyspaceSelectValuesRef = [];
+    let keyspaceRef = "";
 
-  _handleDcFilterSuggestions(textInputValue, possibleSuggestionsArray) {
-    var lowerCaseQuery = textInputValue.toLowerCase();
-    let datacenters = this.state.datacenters;
- 
-    return possibleSuggestionsArray.filter(function(suggestion)  {
-        return suggestion.toLowerCase().includes(lowerCaseQuery) && $.inArray(suggestion, datacenters.split(','))==-1;
-    })
-  },
+    keyspaceListRef.length = 0;
 
-    // Blacklist tag list functions
-    _handleBlacklistAddition(table) {
-      if (this.state.tableList.length == 0 && table.length > 1) {
-        let blacklist = this.state.blacklistList;
-        if ($.inArray(table, this.state.blacklistedTables.split(','))==-1) {
-          blacklist.push({
-              id: this._create_UUID(),
-              text: table
-          });
-          this.setState({blacklistList: blacklist, blacklistedTables: blacklist.map(table => table.text).join(',')});
-          this._checkValidity();
-          this.setState({tablelistReadOnly: true});
-        }
-      }
-    },
-  
-    _handleBlacklistDelete(i) {
-        let blacklist = this.state.blacklistList;
-        blacklist.splice(i, 1);
-        this.setState({blacklistList: blacklist, blacklistedTables: blacklist.map(table => table.text).join(',')});
-        this._checkValidity();
-        this.setState({tablelistReadOnly: (blacklist.length>0)});
-    },
-  
-    _handleBlacklistFilterSuggestions(textInputValue, possibleSuggestionsArray) {
-      var lowerCaseQuery = textInputValue.toLowerCase();
-      let blacklist = this.state.blacklistedTables;
-      let tables = this.state.tables;
-   
-      return possibleSuggestionsArray.filter(function(suggestion)  {
-          return suggestion.toLowerCase().includes(lowerCaseQuery) && $.inArray(suggestion, blacklist.split(','))==-1
-                 && $.inArray(suggestion, tables.split(','))==-1;
-      })
-    },
-  
-    // Tables tag list functions
-    _handleTableAddition(table) {
-      if (this.state.blacklistList.length == 0 && table.length > 1) { 
-        let tables = this.state.tableList;
-        if ($.inArray(table, this.state.tables.split(','))==-1) {
-          tables.push({
-              id: this._create_UUID(),
-              text: table
-          });
-          this.setState({tableList: tables, tables: tables.map(table => table.text).join(',')});
-          this._checkValidity();
-          this.setState({blacklistReadOnly: true});
-        }
-      }
-    },
-  
-    _handleTableDelete(i) {
-        let tables = this.state.tableList;
-        tables.splice(i, 1);
-        this.setState({tableList: tables, tables: tables.map(table => table.text).join(',')});
-        this._checkValidity();
-        this.setState({blacklistReadOnly: (tables.length > 0)});
-    },
-  
-    // Keyspace tag list functions
-    _handleKeyspaceAddition(ks) {
-      let keyspaces = this.state.keyspaceList;
-      if (keyspaces.length==0) {
-        if ($.inArray(ks, this.state.keyspace.split(','))==-1) {
-          keyspaces.push({
-              id: this._create_UUID(),
-              text: ks
-          });
-          this.setState({keyspaceList: keyspaces, keyspace: ks, keyspaces: keyspaces.map(ks => ks.text).join(',')});
-          this._checkValidity();
-          this._getTableSuggestions(ks);
-        }
-      }
-    },
-  
-    _handleKeyspaceDelete(i) {
-        let keyspaces = this.state.keyspaceList;
-        keyspaces.splice(i, 1);
-        this.setState({keyspaceList: keyspaces, keyspace: "", keyspaces: keyspaces.map(ks => ks.text).join(',')});
-        this._checkValidity();
-        this._getTableSuggestions("");
-    },
-
-    _handleKeyspaceFilterSuggestions(textInputValue, possibleSuggestionsArray) {
-      var lowerCaseQuery = textInputValue.toLowerCase();
-      let keyspaces = this.state.keyspaceList;
-   
-      return possibleSuggestionsArray.filter(function(suggestion)  {
-          return suggestion.toLowerCase().includes(lowerCaseQuery) && keyspaces.length==0;
-      })
-    },
-
-    _create_UUID(){
-      var dt = new Date().getTime();
-      var uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-          var r = (dt + Math.random()*16)%16 | 0;
-          dt = Math.floor(dt/16);
-          return (c=='x' ? r :(r&0x3|0x8)).toString(16);
+    if (valueContext) {
+      keyspaceListRef.push({
+        id: this._create_UUID(),
+        text: valueContext.value
       });
-      return uuid;
-    },
+      keyspaceRef = valueContext.value;
+      keyspaceSelectValuesRef.push({
+        label: valueContext.value, value: valueContext.value
+      });
+    }
+
+    this.setState({
+      keyspace: keyspaceRef,
+      keyspaceList: keyspaceListRef,
+      keyspaceSelectValues: keyspaceSelectValuesRef,
+    });
+    this._checkValidity();
+    this._handleLinkedSelectOnChange(null, {name: "in_tables"});
+    this._handleLinkedSelectOnChange(null, {name: "in_blacklistedTables"});
+    this._getTableOptions(keyspaceRef);
+  },
+
+  _create_UUID(){
+    var dt = new Date().getTime();
+    var uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
+        var r = (dt + Math.random()*16)%16 | 0;
+        dt = Math.floor(dt/16);
+        return (c === 'x' ? r : (r&0x3|0x8)).toString(16);
+    });
+    return uuid;
+  },
 
   render: function() {
 
@@ -343,9 +364,32 @@ const repairForm = React.createClass({
       addMsg = <div className="alert alert-danger" role="alert">{this.state.addRepairResultMsg}</div>
     }
 
-    const clusterItems = this.state.clusterNames.sort().map(name =>
-      <option key={name} value={name}>{name}</option>
-    );
+    const clusterNameOptions = this.state.clusterNames.sort().map(name => {
+      return {label: name, value: name};
+    });
+
+    const clusterNamePlaceholder = this.state.clusterName ? this.state.clusterName : "Select cluster";
+    const ownerPlaceholder = `owner name for the ${this.state.formType} run (any string)`;
+
+    let formHeaderLabel = "Start a new repair";
+    let repairButtonLabel = "Repair";
+    let repairButtonClassName = "btn btn-warning";
+
+    if (this.state.formType === "schedule") {
+      formHeaderLabel = "Add a new schedule";
+      repairButtonLabel = "Add Schedule";
+      repairButtonClassName = "btn btn-success";
+    }
+
+    switch (this.state.formType) {
+      case "repair":
+        formHeaderLabel = "Start a new repair";
+
+        break;
+      case "schedule":
+        formHeaderLabel = "Add a new schedule";
+        break;
+    }
 
     let advancedMenuDownStyle = {
       display: "inline-block" 
@@ -355,7 +399,7 @@ const repairForm = React.createClass({
       display: "none" 
     }
 
-    if(this.state.advancedFormCollapsed == false) {
+    if (this.state.advancedFormCollapsed == false) {
       advancedMenuDownStyle = {
         display: "none"
       }
@@ -364,42 +408,90 @@ const repairForm = React.createClass({
       }
     }
 
+    let customInput = "";
+    if (this.state.formType === "repair") {
+      customInput = (
+        <div className="form-group">
+          <label htmlFor="in_cause" className="col-sm-3 control-label">Cause</label>
+          <div className="col-sm-9 col-md-7 col-lg-5">
+            <input type="text" className="form-control" value={this.state.cause}
+              onChange={this._handleChange} id="in_cause" placeholder="reason repair was started"/>
+          </div>
+        </div>
+      );
+    }
+    else if (this.state.formType === "schedule") {
+      customInput = (
+        <div>
+        <div className="form-group">
+          <label htmlFor="in_startTime" className="col-sm-3 control-label">Start time*</label>
+          <div className="col-sm-9 col-md-7 col-lg-5">
+            <DatePicker
+              selected={this.state.startTime}
+              onChange={value => this.setState({startTime: value})}
+              showTimeSelect
+              timeFormat="HH:mm"
+              timeIntervals={15}
+              timeCaption="Time"
+              dateFormat="d MMMM yyyy HH:mm"
+            />
+          </div>
+        </div>
+        <div className="form-group">
+          <label htmlFor="in_intervalDays" className="col-sm-3 control-label">Interval in days*</label>
+          <div className="col-sm-9 col-md-7 col-lg-5">
+            <input type="number" required className="form-control" value={this.state.intervalDays}
+              onChange={this._handleChange} id="in_intervalDays" placeholder="amount of days to wait between scheduling new repairs, (e.g. 7 for weekly)"/>
+          </div>
+        </div>
+        </div>
+      );
+    }
+
     const keyspaceInputStyle = this.state.keyspaceList.length > 0 ? 'form-control-hidden':'form-control';
 
-    const advancedSettingsHeader = <div className="panel-title" >
-    <a href="#advanced-form" data-toggle="collapse" onClick={this._toggleAdvancedSettingsDisplay}>Advanced settings
-    &nbsp; <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={advancedMenuDownStyle}></span>
-           <span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={advancedMenuUpStyle}></span></a></div>
+    const advancedSettingsHeader = (
+      <div className="panel-title" >
+        <a href="#advanced-form" data-toggle="collapse" onClick={this._toggleAdvancedSettingsDisplay}>
+          Advanced settings&nbsp;
+          <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={advancedMenuDownStyle}></span>
+          <span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={advancedMenuUpStyle}></span>
+        </a>
+      </div>
+    );
 
-
-    const form = <div className="row">
+    const form = (
+      <div className="row">
         <div className="col-lg-12">
-
           <form className="form-horizontal form-condensed">
-
             <div className="form-group">
               <label htmlFor="in_clusterName" className="col-sm-3 control-label">Cluster*</label>
               <div className="col-sm-9 col-md-7 col-lg-5">
-                <select className="form-control" id="in_clusterName"
-                  onChange={this._handleChange} value={this.state.clusterName}>
-                  {clusterItems}
-                </select>
+                <Select
+                  id="in_clusterName"
+                  name="in_clusterName"
+                  classNamePrefix="select"
+                  options={clusterNameOptions}
+                  placeholder={clusterNamePlaceholder}
+                  onChange={this._handleSelectOnChange}
+                />
               </div>
             </div>
 
             <div className="form-group">
             <label htmlFor="in_keyspace" className="col-sm-3 control-label">Keyspace*</label>
-            <div className="col-sm-9 col-md-7 col-lg-5">
-              <ReactTags id={'in_keyspace'} tags={this.state.keyspaceList}
-                suggestions={this.state.keyspaceSuggestions}
-                labelField={'text'} handleAddition={this._handleKeyspaceAddition} 
-                handleInputBlur={this._handleKeyspaceAddition} 
-                handleDelete={this._handleKeyspaceDelete}
-                placeholder={'Add a keyspace'}
-                handleFilterSuggestions={this._handleKeyspaceFilterSuggestions}
-                classNames={{
-                    tagInputField: keyspaceInputStyle
-                  }}/>
+              <div className="col-sm-9 col-md-7 col-lg-5">
+                <Select
+                  id="in_keyspace"
+                  name="in_keyspace"
+                  classNamePrefix="select"
+                  isClearable
+                  isSearchable
+                  options={this.state.keyspaceOptions}
+                  value={this.state.keyspaceSelectValues}
+                  placeholder="Add a keyspace"
+                  onChange={this._handleKeyspaceSelectOnChange}
+                />
               </div>
             </div>
             
@@ -407,19 +499,10 @@ const repairForm = React.createClass({
               <label htmlFor="in_owner" className="col-sm-3 control-label">Owner*</label>
               <div className="col-sm-9 col-md-7 col-lg-5">
                 <input type="text" required className="form-control" value={this.state.owner}
-                  onChange={this._handleChange} id="in_owner" placeholder="owner name for the repair run (any string)"/>
+                  onChange={this._handleChange} id="in_owner" placeholder={ownerPlaceholder}/>
               </div>
             </div>
- 
-
-            <div className="form-group">
-              <label htmlFor="in_cause" className="col-sm-3 control-label">Cause</label>
-              <div className="col-sm-9 col-md-7 col-lg-5">
-                <input type="text" className="form-control" value={this.state.cause}
-                  onChange={this._handleChange} id="in_cause" placeholder="reason repair was started"/>
-              </div>
-            </div>
-
+            {customInput}
             <div className="form-group">
               <div className="col-sm-offset-1 col-sm-9">
                 <div className="panel panel-info">
@@ -429,64 +512,74 @@ const repairForm = React.createClass({
                   <div className="panel-body collapse" id="advanced-form">
                     <div className="form-group">
                     <label htmlFor="in_tables" className="col-sm-3 control-label">Tables</label>
-                    <div className="col-sm-14 col-md-12 col-lg-9">
-                      <ReactTags id={'in_tables'} tags={this.state.tableList}
-                        suggestions={this.state.tableSuggestions}
-                        labelField={'text'} handleAddition={this._handleTableAddition} 
-                        handleInputBlur={this._handleTableAddition} 
-                        handleDelete={this._handleTableDelete}
-                        readOnly={this.state.tablelistReadOnly}
-                        placeholder={'Add a table (optional)'}
-                        handleFilterSuggestions={this._handleBlacklistFilterSuggestions}
-                        classNames={{
-                            tagInputField: 'form-control'
-                          }}/>
+                      <div className="col-sm-14 col-md-12 col-lg-9">
+                        <Select
+                          id="in_tables"
+                          name="in_tables"
+                          classNamePrefix="select"
+                          isClearable
+                          isSearchable
+                          isMulti
+                          isDisabled={this.state.tablesSelectDisabled}
+                          options={this.state.tableOptions}
+                          value={this.state.tablesSelectValues}
+                          placeholder="Add a table (optional)"
+                          onChange={this._handleLinkedSelectOnChange}
+                        />
                       </div>
                     </div>
                     <div className="form-group">
                     <label htmlFor="in_blacklist" className="col-sm-3 control-label">Blacklist</label>
-                    <div className="col-sm-14 col-md-12 col-lg-9">
-                      <ReactTags id={'in_blacklist'} tags={this.state.blacklistList}
-                        suggestions={this.state.tableSuggestions}
-                        labelField={'text'} handleAddition={this._handleBlacklistAddition} 
-                        handleInputBlur={this._handleBlacklistAddition}
-                        handleDelete={this._handleBlacklistDelete}
-                        readOnly={this.state.blacklistReadOnly}
-                        placeholder={'Add a table (optional)'}
-                        handleFilterSuggestions={this._handleBlacklistFilterSuggestions}
-                        classNames={{
-                            tagInputField: 'form-control'
-                          }}/>
+                      <div className="col-sm-14 col-md-12 col-lg-9">
+                        <Select
+                          id="in_blacklistedTables"
+                          name="in_blacklistedTables"
+                          classNamePrefix="select"
+                          isClearable
+                          isSearchable
+                          isMulti
+                          isDisabled={this.state.blacklistedTablesSelectDisabled}
+                          options={this.state.tableOptions}
+                          value={this.state.blacklistedTablesSelectValues}
+                          placeholder="Add a table (optional)"
+                          onChange={this._handleLinkedSelectOnChange}
+                        />
                       </div>
                     </div>
                     <div className="form-group">
                       <label htmlFor="in_nodes" className="col-sm-3 control-label">Nodes</label>
                       <div className="col-sm-14 col-md-12 col-lg-9">
-                      <ReactTags id={'in_nodes'} tags={this.state.nodeList}
-                        suggestions={this.state.nodeSuggestions}
-                        labelField={'text'} handleAddition={this._handleAddition} 
-                        handleInputBlur={this._handleAddition} 
-                        handleDelete={this._handleDelete}
-                        placeholder={'Add a node (optional)'}
-                        handleFilterSuggestions={this._handleNodeFilterSuggestions}
-                        classNames={{
-                            tagInputField: 'form-control'
-                          }}/>
+                        <Select
+                          id="in_nodes"
+                          name="in_nodes"
+                          classNamePrefix="select"
+                          isClearable
+                          isSearchable
+                          isMulti
+                          isDisabled={this.state.nodesSelectDisabled}
+                          options={this.state.nodeOptions}
+                          value={this.state.nodesSelectValues}
+                          placeholder="Add a node (optional)"
+                          onChange={this._handleLinkedSelectOnChange}
+                        />
                       </div>
                     </div>
                     <div className="form-group">
                       <label htmlFor="in_datacenters" className="col-sm-3 control-label">Datacenters</label>
                       <div className="col-sm-14 col-md-12 col-lg-9">
-                      <ReactTags id={'in_datacenters'} tags={this.state.datacenterList}
-                      suggestions={this.state.datacenterSuggestions}
-                      labelField={'text'} handleAddition={this._handleDcAddition} 
-                      handleInputBlur={this._handleDcAddition} 
-                      handleDelete={this._handleDcDelete}
-                      placeholder={'Add a datacenter (optional)'}
-                      handleFilterSuggestions={this._handleDcFilterSuggestions}
-                      classNames={{
-                          tagInputField: 'form-control'
-                        }}/>
+                        <Select
+                          id="in_datacenters"
+                          name="in_datacenters"
+                          classNamePrefix="select"
+                          isClearable
+                          isSearchable
+                          isMulti
+                          isDisabled={this.state.datacentersSelectDisabled}
+                          options={this.state.datacenterOptions}
+                          value={this.state.datacentersSelectValues}
+                          placeholder="Add a datacenter (optional)"
+                          onChange={this._handleLinkedSelectOnChange}
+                        />
                       </div>
                     </div>
                     <div className="form-group">
@@ -499,13 +592,18 @@ const repairForm = React.createClass({
                     <div className="form-group">
                       <label htmlFor="in_parallelism" className="col-sm-3 control-label">Parallelism</label>
                       <div className="col-sm-14 col-md-12 col-lg-9">
-                        <select className="form-control" id="in_parallelism"
-                          onChange={this._handleChange} value={this.state.parallelism}>
-                          <option value=""></option>
-                          <option value="SEQUENTIAL">Sequential</option>
-                          <option value="PARALLEL">Parallel</option>
-                          <option value="DATACENTER_AWARE">DC-Aware</option>
-                        </select>
+                        <Select
+                          id="in_parallelism"
+                          name="in_parallelism"
+                          classNamePrefix="select"
+                          options={[
+                            {label: "Sequential", value: "SEQUENTIAL"},
+                            {label: "Parallel", value: "PARALLEL"},
+                            {label: "DC-Aware", value: "DATACENTER_AWARE"},
+                          ]}
+                          placeholder="Select parallelism"
+                          onChange={this._handleSelectOnChange}
+                        />
                       </div>
                     </div>
                     <div className="form-group">
@@ -519,11 +617,14 @@ const repairForm = React.createClass({
                     <div className="form-group">
                       <label htmlFor="in_incrementalRepair" className="col-sm-3 control-label">Incremental</label>
                       <div className="col-sm-14 col-md-12 col-lg-9">
-                        <select className="form-control" id="in_incrementalRepair"
-                          onChange={this._handleChange} value={this.state.incrementalRepair}>
-                          <option value="false">false</option>
-                          <option value="true">true</option>
-                        </select>
+                        <Select
+                          id="in_incrementalRepair"
+                          name="in_incrementalRepair"
+                          classNamePrefix="select"
+                          options={[{label: "true", value: "true"}, {label: "false", value: "false"}]}
+                          placeholder="false"
+                          onChange={this._handleSelectOnChange}
+                        />
                       </div>
                     </div>
                     <div className="form-group">
@@ -540,16 +641,14 @@ const repairForm = React.createClass({
             </div>
             <div className="form-group">
               <div className="col-sm-offset-3 col-sm-9">
-                <button type="button" className="btn btn-warning" disabled={!this.state.submitEnabled}
-                  onClick={this._onAdd}>Repair</button>
+                <button type="button" className={repairButtonClassName} disabled={!this.state.submitEnabled}
+                  onClick={this._onAdd}>{repairButtonLabel}</button>
               </div>
             </div>            
           </form>
-
+        </div>
       </div>
-    </div>
-
-    
+    );
 
     let menuDownStyle = {
       display: "inline-block" 
@@ -568,21 +667,27 @@ const repairForm = React.createClass({
       }
     }
 
-    const formHeader = <div className="panel-title" ><a href="#repair-form" data-toggle="collapse" onClick={this._toggleFormDisplay}>Start a new repair&nbsp; <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={menuDownStyle}></span><span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={menuUpStyle}></span></a></div>
+    const formHeader = (
+      <div className="panel-title" >
+        <a href="#repair-form" data-toggle="collapse" onClick={this._toggleFormDisplay}>
+          {formHeaderLabel}&nbsp;
+          <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={menuDownStyle}></span>
+          <span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={menuUpStyle}></span>
+        </a>
+      </div>
+    );
 
-
-
-
-
-    return (<div className="panel panel-warning">
-              <div className="panel-heading">
-                {formHeader}
-              </div>
-              <div className="panel-body collapse" id="repair-form">
-                {addMsg}
-                {form}
-              </div>
-            </div>);
+    return (
+      <div className="panel panel-warning">
+        <div className="panel-heading">
+          {formHeader}
+        </div>
+        <div className="panel-body collapse" id="repair-form">
+          {addMsg}
+          {form}
+        </div>
+      </div>
+    );
   }
 });
 
