@@ -15,6 +15,9 @@
 //  limitations under the License.
 
 import React from "react";
+import CreateReactClass from 'create-react-class';
+import PropTypes from 'prop-types';
+import Select from 'react-select';
 import moment from "moment";
 import {RowDeleteMixin, RowAbortMixin, StatusUpdateMixin, DeleteStatusMessageMixin, CFsListRender, CFsCountListRender} from "jsx/mixin";
 import ProgressBar from 'react-bootstrap/lib/ProgressBar';
@@ -22,7 +25,7 @@ import Button from 'react-bootstrap/lib/Button';
 import Modal from 'react-bootstrap/lib/Modal';
 import segmentList from 'jsx/segment-list'
 
-const TableRow = React.createClass({
+const TableRow = CreateReactClass({
   mixins: [RowDeleteMixin, StatusUpdateMixin, RowAbortMixin],
 
   _viewSegments: function(id) {
@@ -31,7 +34,7 @@ const TableRow = React.createClass({
   },
 
   segmentsButton: function(id) {
-    return <Button bsSize="xs" bsStyle="info" onClick={() => this._viewSegments(id)}>View segments</Button>
+    return <button className="btn btn-xs btn-info" onClick={() => this._viewSegments(id)}>View segments</button>
   },
 
   render: function() {
@@ -64,15 +67,21 @@ const TableRow = React.createClass({
       progressStyleColor = "danger";
     }
 
-
-    
-    const btnStartStop = this.props.row.state == 'ABORTED' ? null : this.statusUpdateButton();
-    const btnAbort = state == 'RUNNING' || state == 'PAUSED' ? this.abortButton() : this.deleteButton();
+    const btnStartStop = (this.props.row.state === 'ABORTED' ? null : this.statusUpdateButton());
+    const btnAbort = (state === 'RUNNING') || (state === 'PAUSED' ? this.abortButton() : this.deleteButton());
     const btnSegment = this.segmentsButton(this.props.row.id);
-    const active = state == 'RUNNING';
-    let repairProgress = <ProgressBar now={Math.round((segsRepaired*100)/segsTotal)} active={active} bsStyle={progressStyleColor} 
-                   label={segsRepaired + '/' + segsTotal}
-                   key={progressID}/>
+    const progressNow = Math.round((segsRepaired * 100) / segsTotal)
+    const progressLabel = `${segsRepaired} / ${segsTotal}`
+
+    const repairProgress = (
+      <ProgressBar
+        now={progressNow}
+        active={state === 'RUNNING' ? 1 : 0}
+        bsstyle={progressStyleColor}
+        label={progressLabel}
+      />
+    );
+
 
     return (
     <tr>
@@ -98,7 +107,7 @@ const TableRow = React.createClass({
 
 });
 
-const TableRowDetails = React.createClass({
+const TableRowDetails = CreateReactClass({
   getInitialState() {
       return { intensity: this.props.row.intensity };
   },
@@ -254,31 +263,39 @@ const TableRowDetails = React.createClass({
 
 });
 
-const repairList = React.createClass({
+const repairList = CreateReactClass({
   mixins: [DeleteStatusMessageMixin],
 
   propTypes: {
-    repairs: React.PropTypes.object.isRequired,
-    clusterNames: React.PropTypes.object.isRequired,
-    deleteSubject: React.PropTypes.object.isRequired,
-    deleteResult: React.PropTypes.object.isRequired,
-    updateStatusSubject: React.PropTypes.object.isRequired,
-    updateIntensitySubject: React.PropTypes.object.isRequired,
-    currentCluster: React.PropTypes.string.isRequired,
-    changeCurrentCluster: React.PropTypes.func.isRequired
-
+    repairs: PropTypes.object.isRequired,
+    clusterNames: PropTypes.object.isRequired,
+    deleteSubject: PropTypes.object.isRequired,
+    deleteResult: PropTypes.object.isRequired,
+    updateStatusSubject: PropTypes.object.isRequired,
+    updateIntensitySubject: PropTypes.object.isRequired,
+    currentCluster: PropTypes.string.isRequired,
+    changeCurrentCluster: PropTypes.func.isRequired
   },
 
   getInitialState: function() {
-    return {repairs: null, deleteResultMsg: null, clusterNames:[],
-      currentCluster:this.props.currentCluster, 
-      runningCollapsed: false, doneCollapsed: false,
-      modalShow: false, repairRunId: '',
-      height: 0, width: 0,
-      numberOfElementsToDisplay: 10};
+    return {
+      repairs: null,
+      deleteResultMsg: null,
+      clusterNames: [],
+      currentCluster: this.props.currentCluster,
+      currentClusterSelectValue: {value: 'all', label: 'all'},
+      runningCollapsed: false,
+      doneCollapsed: false,
+      modalShow: false,
+      repairRunId: '',
+      height: 0,
+      width: 0,
+      numberOfElementsToDisplay: 10,
+      numberOfElementsToDisplaySelectValue: {value: 10, label: '10'},
+    };
   },
 
-  componentWillMount: function() {
+  UNSAFE_componentWillMount: function() {
     this._clusterNamesSubscription = this.props.clusterNames.subscribeOnNext(obs =>
       obs.subscribeOnNext(names => this.setState({clusterNames: names}))
     );
@@ -314,20 +331,24 @@ const repairList = React.createClass({
     this.setState({ width: window.innerWidth, height: window.innerHeight });
   },
 
-  _handleChange: function(e) {
-    var v = e.target.value;
-    var n = e.target.id.substring(3); // strip in_ prefix
+  _handleSelectOnChange: function(valueContext, actionContext) {
+    const nameRef = actionContext.name.split("_")[1];
+    const nameSelectValueRef = `${nameRef}SelectValue`;
 
-    // update state
-    const state = this.state;
-    state[n] = v;
-    this.replaceState(state);
+    let newSelectValue = {};
+    let newValueRef = "";
 
-    // validate
-    const valid = state.currentCluster;
-    this.setState({submitEnabled: valid});
+    if (valueContext) {
+        newSelectValue = valueContext;
+        newValueRef = valueContext.value;
+    }
+
+    let newState = {};
+    newState[nameRef] = newValueRef;
+    newState[nameSelectValueRef] = newSelectValue;
+
+    this.setState(newState);
     this.props.changeCurrentCluster(this.state.currentCluster);
-    console.log("changed cluster to " + this.state.currentCluster);
   },
 
   _toggleRunningDisplay: function() {
@@ -374,19 +395,32 @@ const repairList = React.createClass({
       return 0;
     }
 
-    let modalClose = () => this.setState({ modalShow:false, repairRunId: ''});
+    const modalClose = () => this.setState({ modalShow: false, repairRunId: ''});
 
-    const segmentModal = <SegmentModal show={this.state.modalShow} onHide={modalClose} repairRunId={this.state.repairRunId} height={this.state.height} width={this.state.width}/>;
+    const segmentModal = (
+      <Modal show={this.state.modalShow} onHide={modalClose} height={this.state.height} width={this.state.width} size="lg" aria-labelledby="contained-modal-title-lg" dialogClassName="large-modal">
+        <Modal.Header closeButton>
+          <Modal.Title id="contained-modal-title-lg">Segments</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <iframe src={"segments.html?repairRunId="+this.state.repairRunId} width="100%" height={parseInt(this.state.height)-200} frameBorder="0"/>
+        </Modal.Body>
+        <Modal.Footer>
+          <Button onClick={modalClose}>Close</Button>
+        </Modal.Footer>
+      </Modal>
+    );
+
     if (this.state.repairs !== null) {
         rowsRunning = this.state.repairs.sort(compareStartTimeReverse)
             .filter(repair => this.state.currentCluster === "all" || this.state.currentCluster === repair.cluster_name)
             .filter(repair => (repair.state === "RUNNING" || repair.state === "PAUSED" || repair.state === "NOT_STARTED"))
             .map(repair =>
                 <tbody key={repair.id + '-rows'}>
-                    <TableRow row={repair} key={repair.id + '-head'}
+                    <TableRow row={repair}
                               deleteSubject={this.props.deleteSubject}
                               updateStatusSubject={this.props.updateStatusSubject} showSegments={this._displaySegments}/>
-                    <TableRowDetails row={repair} key={repair.id + '-details'}
+                    <TableRowDetails row={repair}
                                      updateIntensitySubject={this.props.updateIntensitySubject}/>
                 </tbody>
             );
@@ -397,37 +431,48 @@ const repairList = React.createClass({
             .slice(0, this.state.numberOfElementsToDisplay)
             .map(repair =>
                 <tbody key={repair.id + '-rows'}>
-                    <TableRow row={repair} key={repair.id + '-head'}
+                    <TableRow row={repair}
                               deleteSubject={this.props.deleteSubject}
                               updateStatusSubject={this.props.updateStatusSubject} showSegments={this._displaySegments}/>
-                    <TableRowDetails row={repair} key={repair.id + '-details'}/>
+                    <TableRowDetails row={repair}/>
                 </tbody>
             );
     }
 
-    const clusterItems = this.state.clusterNames.sort().map(name =>
-      <option key={name} value={name}>{name}</option>
-    );
+    let selectClusterItems = this.state.clusterNames.sort().map(name => {
+        { return { value: name, label: name}; }
+    });
+
+    selectClusterItems.unshift({value: 'all', label: 'all'});
 
     const clusterFilter = <form className="form-horizontal form-condensed">
             <div className="form-group">
               <label htmlFor="in_currentCluster" className="col-sm-3 control-label">Filter cluster:</label>
               <div className="col-sm-7 col-md-5 col-lg-3">
-                <select className="form-control" id="in_currentCluster"
-                  onChange={this._handleChange} value={this.state.currentCluster}>
-                  <option key="all" value="all">All</option>
-                  {clusterItems}
-                </select>
+                <Select
+                    id="in_currentCluster"
+                    name="in_currentCluster"
+                    classNamePrefix="select"
+                    options={selectClusterItems}
+                    value={this.state.currentClusterSelectValue}
+                    onChange={this._handleSelectOnChange}
+                />
               </div>
               <label htmlFor="in_numberOfElementsToDisplay" className="col-sm-1 control-label">Display:</label>
               <div className="col-sm-5 col-md-3 col-lg-1">
-                <select className="form-control" id="in_numberOfElementsToDisplay"
-                  onChange={this._handleChange} value={this.state.numberOfElementsToDisplay}>
-                  <option key="10" value="10">Last 10</option>
-                  <option key="25" value="25">Last 25</option>
-                  <option key="50" value="50">Last 50</option>
-                  <option key="100" value="100">Last 100</option>
-                </select>
+                <Select
+                    id="in_numberOfElementsToDisplay"
+                    name="in_numberOfElementsToDisplay"
+                    classNamePrefix="select"
+                    options={[
+                        {value: 10, label: "10"},
+                        {value: 25, label: "25"},
+                        {value: 50, label: "50"},
+                        {value: 100, label: "100"},
+                    ]}
+                    value={this.state.numberOfElementsToDisplaySelectValue}
+                    onChange={this._handleSelectOnChange}
+                />
               </div>
             </div>
     </form>
@@ -528,8 +573,6 @@ const repairList = React.createClass({
     const runningHeader = <div className="panel-title"><a href="#repairs-running" data-toggle="collapse" onClick={this._toggleRunningDisplay}>Running&nbsp; <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={menuRunningDownStyle}></span><span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={menuRunningUpStyle}></span></a></div>
     const doneHeader = <div className="panel-title"><a href="#repairs-done" data-toggle="collapse" onClick={this._toggleDoneDisplay}>Done&nbsp; <span className="glyphicon glyphicon-menu-down" aria-hidden="true" style={menuDoneDownStyle}></span><span className="glyphicon glyphicon-menu-up" aria-hidden="true" style={menuDoneUpStyle}></span></a></div>
 
-
-
     return (
             <div>
               {segmentModal}
@@ -552,27 +595,6 @@ const repairList = React.createClass({
               </div>
             </div>);
   }
-});
-
-const SegmentModal = React.createClass({
-  getInitialState: function() {
-    return {height: this.props.height};
-  },
-  render() {
-    return (
-      <Modal  {...this.props} bsSize="large" aria-labelledby="contained-modal-title-lg" dialogClassName="large-modal">
-        <Modal.Header closeButton>
-          <Modal.Title id="contained-modal-title-lg">Segments</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <iframe src={"segments.html?repairRunId="+this.props.repairRunId} width="100%" height={parseInt(this.props.height)-200} frameBorder="0"/>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button onClick={this.props.onHide}>Close</Button>
-        </Modal.Footer>
-      </Modal>
-    );
-  },
 });
 
 
