@@ -19,17 +19,34 @@ var webpack = require('webpack');
 var HtmlWebpackPlugin = require('html-webpack-plugin');
 
 var _commonDeps = [
-  "sb-admin-2.css", "react-widgets.css", "bootstrap.css", "theme.css",
-  "timeline.css", "font-awesome.css", "metisMenu.css", "style.scss",
-  "jquery", "react", "bootstrap", "metisMenu", "sb-admin-2", "rxjs", "bootstrap-multiselect.css", "datatables.net-bs.css"
+  "sb-admin-2.css",
+  "bootstrap.css",
+  "theme.css",
+  "timeline.css",
+  "font-awesome.css",
+  "metisMenu.css",
+  "style.scss",
+  "jquery",
+  "react",
+  "bootstrap",
+  "metisMenu",
+  "sb-admin-2",
+  "rxjs",
+  "bootstrap-multiselect.css",
+  "datatables.net-bs.css",
+  "react-datepicker.css"
 ];
 
 // include hot reload deps in dev mode
-const isDev = process.argv.indexOf('-d') !== -1 || process.env.BUILD_DEV;
-if(isDev) {
+const isDev = process.argv.indexOf('-d') !== -1 || !!+process.env.BUILD_DEV;
+if (isDev) {
+  console.log('Running in development mode!');
   _commonDeps.push("webpack-dev-server/client?http://0.0.0.0:8000"); // WebpackDevServer host and port
   _commonDeps.push("webpack/hot/only-dev-server");
 }
+
+const reaperHost = process.env.REAPER_HOST || '127.0.0.1';
+console.log(`Reaper remote host is '${reaperHost}'`);
 
 module.exports = {
   entry: {
@@ -68,7 +85,6 @@ module.exports = {
     ],
     alias: {
       "jquery": "jquery/dist/jquery",
-      "react-widgets.css": "react-widgets/dist/css/react-widgets.css",
       "bootstrap.css": "bootstrap.css",
       "theme.css": "bootstrap-theme.css",
       "sb-admin-2.css": "startbootstrap-sb-admin-2/dist/css/sb-admin-2.css",
@@ -81,7 +97,8 @@ module.exports = {
       "datatables.net": "datatables.net/js/jquery.dataTables.js",
       "datatables.net-bs.js": "datatables.net-bs/js/dataTables.bootstrap.js",
       "datatables.net-bs.css": "datatables.net-bs/css/dataTables.bootstrap.css",
-      "bootstrap-multiselect.css": "react-boostrap-multiselect/css/bootstrap-multiselect.css"
+      "bootstrap-multiselect.css": "react-boostrap-multiselect/css/bootstrap-multiselect.css",
+      "react-datepicker.css": "react-datepicker/dist/react-datepicker.css"
     },
     extensions: ['.js', '.jsx']
   },
@@ -159,6 +176,10 @@ module.exports = {
     }),
     new webpack.HotModuleReplacementPlugin(),
     new webpack.NoEmitOnErrorsPlugin(),
+    new webpack.DefinePlugin({
+      'GLOBAL_REAPER_HOST': JSON.stringify(reaperHost),
+      'GLOBAL_IS_DEV': isDev,
+    }),
   ],
   optimization: {
     splitChunks: {
@@ -169,13 +190,14 @@ module.exports = {
       }
     }
   },
+  mode: isDev ? "development" : "production",
   module: {
     rules:[
       {
         test: /\.(js|jsx)$/,
         exclude: /(node_modules|bower_components)/,
         use: [
-          { loader: "react-hot-loader"},
+          { loader: "react-hot-loader/webpack" },
           {
             loader:"babel-loader",
             options: {
